creation of repositories screen

+++ data/repo/src/main/java/com/example/github/data/repo/datasource/UserDataSource.kt
@@ -3,4 +3,4 @@ package com.example.github.data.repo.datasource
 import androidx.paging.PagingSource
 import com.example.github.data.repo.model.User
 
-internal abstract class UserDataSource : PagingSource<Int, User>()
+internal abstract class UserDataSource : PagingSource<Long, User>()
